Tarefa 08/requirements.txt
fastapi
uvicorn
jinja2
pandas
scikit-learn
joblib
transformers
torch
google-api-python-client
python-dotenv
nltk
spacy
matplotlib
tqdm
